+++ tests/test-docker-mw-all-dbs.sh
@@ -66,6 +66,11 @@ test_command_success "./bin/mw docker mediawiki install --dbname mysqlwiki --dbt
 test_command_success "./bin/mw docker mediawiki install --dbname postgreswiki --dbtype postgres"
 test_command_success "./bin/mw docker mediawiki install --dbtype sqlite"
 
+# Make sure mediawiki exec works for alternative db name
+test_command_success "./bin/mw docker mediawiki exec -- MW_DB=mysqlwiki composer phpunit tests/phpunit/unit/includes/XmlTest.php | grep 'seconds'"
+# And doesnt work with a non existant name
+test_command_success "./bin/mw docker mediawiki exec -- MW_DB=ddsadsadsaefault composer phpunit tests/phpunit/unit/includes/XmlTest.php | grep 'Unable to find database'"
+
 # Update the hosts file as we used new wiki names
 if ./bin/mw docker hosts writable; then
     test_command_success "./bin/mw docker hosts add"
@@ -85,3 +90,13 @@ test_curl http://mysqlwiki.mediawiki.mwdd.localhost:$PORT "MediaWiki has been in
 
 # Make sure the expected number of services appear
 test_docker_ps_service_count 9
+
+# Try other DB related commands
+test_command_success "./bin/mw docker postgres stop"
+test_command_success "./bin/mw docker postgres start"
+test_command_success "./bin/mw docker postgres exec echo foo"
+test_command_success "./bin/mw docker mysql-replica stop"
+test_command_success "./bin/mw docker mysql-replica start"
+test_command_success "./bin/mw docker mysql-replica exec echo foo"
+# TODO test the mysql and replica "mysql" commands (cli)
+test_command_success "./bin/mw docker mysql exec echo foo"